agent_platform/requirements.txt
fastapi>=0.115.2,<1.0
uvicorn==0.24.0
pydantic==2.5.1
python-multipart
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv>=1.0.1
smolagents==0.1.0
torch
sqlalchemy<1.5,>=1.4.42
databases[sqlite]==0.8.0
gTTS
SpeechRecognition
requests>=2.31.0
pyotp==2.9.0
